Refactored listener tools
