+++ server/templates/linkTemplate/scripts/listenerTools.js
@@ -66,7 +66,7 @@ var providerClick = function (provider, providerUrl) {
     }, 10);
   } else {
     // FYI, this condition can occur when the users preference is 'none', or if the provider they prefer does not stream the song
-    if (provider === 'youtube' || provider === 'itunes' || provider === 'deezer') {
+    if (provider === 'youtube' || provider === 'itunes' || provider === 'deezer' || provider === 'google') {
       window.open(providerUrl, '_blank');
     } else {
       window.location.href = providerUrl;
@@ -80,11 +80,19 @@ var preferenceSave = function (provider) {
   $('.modalEnvironment').css('display', 'none');
   $('body > main').removeClass('blur');
 
-  if (modalSelection.originalPref === 'youtube' || modalSelection.originalPref === 'itunes' || modalSelection.originalPref === 'deezer') {
+  if (modalSelection.originalPref === 'youtube' || modalSelection.originalPref === 'itunes' || modalSelection.originalPref === 'deezer' || modalSelection.originalPref === 'google') {
     window.open(modalSelection.originalUrl, '_blank');
   } else {
     window.location.href = modalSelection.originalUrl;
   }
 };
 
-module.exports = docCookies;
+window.addEventListener('load', function () {
+  var album = document.querySelector('.albumArt');
+  var className = 'unloaded';
+  if (album.classList) {
+    album.classList.remove(className);
+  } else {
+    album.className = album.className.replace(new RegExp('(^|\\b)' + className.split(' ').join('|') + '(\\b|$)', 'gi'), ' ');
+  }
+});
