+++ server/templates/linkTemplate/template.html
@@ -116,7 +116,7 @@
     </div>
   </div>
 
-  <script src="scripts/listenerToolsCopy.js"></script>
+  <script src="scripts/listenerTools.js"></script>
   <script>
 
     window.addEventListener('load', function () {
